[SYCL] Dissolve info_desc_helpers.hpp; fix latent aspect canonical-decl bug
Move the per-class is_*_info_desc<T> SFINAE predicates out of
sycl/detail/info_desc_helpers.hpp into the matching sycl/info/<class>.hpp
headers added in PR #22273. Park is_backend_info_desc<T> and
is_kernel_queue_specific_info_desc<T> in sycl/detail/info_desc_traits.hpp
since neither has a per-class info header. Delete info_desc_helpers.hpp
and repoint helpers consumers (SYCL object headers, source/detail/*_info
headers, ur_info_code.hpp, the three launch_queries e2e tests).

Bug fix: device.hpp and device_selector.hpp carried `enum class aspect;`
as a forward declaration. Clang binds the
[[__sycl_detail__::sycl_type(aspect)]] attribute (and the
__uses_aspects__ parser check in half_type_impl.hpp) to the canonical
declaration. After PR #22273's header reorder, the fwd decl became
canonical when it preceded <sycl/aspects.hpp> in include order, breaking
SYCL device compilation of any TU pulling <sycl/kernel_bundle.hpp> with
-fsycl-device-only. Replace both fwd decls with #include
<sycl/aspects.hpp>; the header is ~30 lines and pulls no STL.

Regression test for the aspect fix:
  clang++ -fsycl -fsycl-device-only -c -include sycl/kernel_bundle.hpp ...
