+++ sycl/source/detail/context_info.hpp
@@ -11,7 +11,7 @@
 #include <detail/context_impl.hpp>
 #include <detail/ur_info_code.hpp>
 #include <sycl/detail/common.hpp>
-#include <sycl/detail/info_desc_helpers.hpp>
+#include <sycl/info/context.hpp>
 #include <sycl/info/info_desc.hpp>
 
 namespace sycl {
