+++ sycl/source/detail/event_info.hpp
@@ -12,7 +12,7 @@
 #include <detail/event_impl.hpp>
 #include <detail/ur_info_code.hpp>
 #include <sycl/detail/common.hpp>
-#include <sycl/detail/info_desc_helpers.hpp>
+#include <sycl/info/event.hpp>
 #include <sycl/info/info_desc.hpp>
 
 namespace sycl {
