+++ sycl/source/detail/ur_info_code.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include <sycl/detail/info_desc_helpers.hpp>
+#include <sycl/detail/info_desc_traits.hpp>
 
 #include <unified-runtime/ur_api.h>
