+++ sycl/test-e2e/Experimental/launch_queries/max_sub_groups.cpp
@@ -2,7 +2,7 @@
 // RUN: %{run} %t.out
 
 #include <sycl/detail/core.hpp>
-#include <sycl/detail/info_desc_helpers.hpp>
+#include <sycl/info/kernel.hpp>
 #include <sycl/kernel.hpp>
 #include <sycl/kernel_bundle.hpp>
 #include <sycl/nd_item.hpp>
